requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
astroid==3.2.4
black==24.8.0
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
dill==0.3.8
distlib==0.3.8
ecdsa==0.19.0
fastapi==0.112.0
filelock==3.15.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
identify==2.6.0
idna==3.7
iniconfig==2.0.0
isort==5.13.2
Mako==1.3.5
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.8.0
psycopg==3.2.1
psycopg2-binary==2.9.9
pyasn1==0.6.0
pydantic==2.8.2
pydantic_core==2.20.1
pylint==3.2.6
pytest==8.3.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
requests-mock==1.12.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
tomlkit==0.13.0
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
virtualenv==20.26.3
